Multiply magnetometer by factory scale factors

+++ regtek.hpp
@@ -10,7 +10,7 @@ namespace regtek {
         const std::array<int16_t, 3>& imu_gyro,
         bool mag_valid,
         bool mag_of,
-        const std::array<int16_t, 3>& mag,
+        const std::array<int32_t, 3>& mag,
         bool alt_valid,
         const uint32_t& alt_p,
         const int32_t& alt_t
